outsource overapproximate methods to module
