+++ src/Algorithms/A20/A20.jl
@@ -1,3 +1,5 @@
+using ..Overapproximate: _convert_or_overapproximate
+
 """
     A20{N} <: AbstractContinuousPost
