+++ src/Algorithms/GLGM06/GLGM06.jl
@@ -1,3 +1,5 @@
+using ..Overapproximate: _convert_or_overapproximate
+
 """
     GLGM06{N, AM, S, D, NG, P, RM} <: AbstractContinuousPost
