+++ src/Discretization/discretization.jl
@@ -7,6 +7,9 @@ include("DiscretizationModule.jl")
 
 include("ApplySetops.jl")
 
+include("Overapproximate.jl")
+@reexport using ..Overapproximate
+
 # Approximation model in discrete time, i.e. without bloating
 using ..DiscretizationModule # TODO Remove
 using ..Exponentiation # TODO Remove
