+++ src/Hybrid/solve.jl
@@ -1,3 +1,5 @@
+using ..Overapproximate: _overapproximate
+
 # =====================================
 # Main solve for hybrid systems
 # =====================================
