+++ test/algorithms/ASB07.jl
@@ -1,4 +1,5 @@
 using ReachabilityAnalysis: IntervalExpAlg
+using ..Overapproximate: _convert_or_overapproximate
 
 @testset "ASB07 algorithm: homogeneous case" begin
     # Example 1 from
